Assigment 4/a4/cs224n-A4-student/requirements.txt
python-dotenv
matplotlib
openai
requests
google-genai
